Move server/etcdserver/txn.go to new package: server/etcdserver/txn
